convertd weight to double, fixd data on Animal.txt
Added translation of Conservation_Status (dispCon()) method in animal class
Error of Number Format Exception fixed
**** To finish Part 1*** 
we need to implement feed() for every specific Animal class

+++ Bird.java
@@ -10,13 +10,13 @@ public class Bird extends Animal
 
 	public int getFlightSpeed(){ return flightSpeed; }
 	public int getWingspan(){ return wingspan; }
-	
-	public Bird(int Age, String Consevervation_status, String Diet, boolean Flight, String Habit, boolean healthy, String Last_Checkup, int Lifespan, String Location, String Name, int Weight, int flightSpeed, int wingspan){
+
+	public Bird(int Age, String Consevervation_status, String Diet, boolean Flight, String Habit, boolean healthy, String Last_Checkup, int Lifespan, String Location, String Name, double Weight, int flightSpeed, int wingspan){
 		super(Age, Consevervation_status, Diet, Flight,  Habit, healthy, Last_Checkup, Lifespan, Location, Name, Weight);
 		this.flightSpeed = flightSpeed;
 		this.wingspan = wingspan;
 	}
-	
+
 	public String toString(){
 		String info = super.toString();
 		info += "Flight speed:" + getFlightSpeed() + " mph.\n";
